requirements.txt
flask
torch
numpy
scipy
pillow
scikit-learn
